Frontend UI

+++ backend/controllers/goalsController.js
@@ -1,10 +1,11 @@
 const asynHandler=require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 //GET
 // uses /api/goal/
 const getGoals =asynHandler(async(req,res)=>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     //To use the line of code below export module.export and set it to an object
     res.status(200).json(goals);
 })
@@ -18,6 +19,7 @@ const setGoal =asynHandler(async(req,res)=>{
     }
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal);
 })
@@ -30,6 +32,17 @@ const updateGoal=asynHandler(async(req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+   // to check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+     // ensuring the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+     res.status(401)
+     throw new Error('User not authorized')
     }
     // creating our updated goal 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.
@@ -48,6 +61,11 @@ const deleteGoal =asynHandler(async(req,res)=>{
         res.status(400)
         throw new Error('Goal not found')
     }
+     // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
    // used delete one remove was not recognized
     await goal.deleteOne();
